feat: implement patient deletion

+++ src/actions/delete-patient/index.ts
@@ -6,6 +6,8 @@ import { actionClient } from "@/lib/next-safe-action";
 import { z } from "zod";
 import { eq } from "drizzle-orm";
 import { revalidatePath } from "next/cache";
+import { auth } from "@/lib/auth";
+import { headers } from "next/headers";
 
 const schema = z.object({
   id: z.string().uuid(),
@@ -14,6 +16,17 @@ const schema = z.object({
 export const deletePatient = actionClient
   .schema(schema)
   .action(async ({ parsedInput }) => {
+    const session = await auth.api.getSession({
+      headers: await headers(),
+    });
+    if (!session?.user) throw new Error("Unauthorized");
+
+    const patient = await db.query.patientsTable.findFirst({
+      where: eq(patientsTable.id, parsedInput.id),
+    });
+    if (!patient || patient.clinicId != session.user.clinic?.id)
+      throw new Error("Paciente não encontrado");
+
     await db.delete(patientsTable).where(eq(patientsTable.id, parsedInput.id));
     revalidatePath("/patients");
   });
